chore: add VBtnGroup vitest file(#20517)
